# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MicroLens Recommendation MVP API",
    "description": "Phase 2D public runtime. Feed orchestration remains explicitly deferred to Phase 4.",
    "version": "0.2.0-phase2d",
    "x-implementation-status": "phase_2d_public_runtime"
  },
  "paths": {
...
    "/api/items/{item_id}": {
      "get": {
        "tags": [
          "feeds"
        ],
        "summary": "Get Item",
        "operationId": "getItem",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-required-roles": [
          "user",
          "operator_readonly",
          "operator",
          "admin"
        ]
      }
    },
    "/api/admin/items": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Search Items",
        "operationId": "searchAdminItems",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "name": "online_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OnlineStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Online Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminItem"
                  },
                  "title": "Response Searchadminitems"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-required-roles": [
          "operator_readonly",
          "operator",
          "admin"
        ]
      }
    },
    "/api/admin/promotions": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Promotion",
        "operationId": "createPromotion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "cookieAuth": [],
            "csrfHeader": []
          }
        ],
        "x-required-roles": [
          "operator",
          "admin"
        ]
      }
    },
    "/api/admin/operation-batches": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Operation Batch",
        "operationId": "createOperationBatch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "cookieAuth": [],
            "csrfHeader": []
          }
        ],
        "x-required-roles": [
          "operator",
          "admin"
        ]
      }
    },
    "/api/admin/operations": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Operations",
        "operationId": "listOperations",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditOperation"
                  },
                  "type": "array",
                  "title": "Response Listoperations"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-required-roles": [
          "operator_readonly",
          "operator",
          "admin"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id",
          "details"
        ],
        "title": "ErrorEnvelope"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemDetailResponse": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "cover": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover"
          },
          "position": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Position"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "item_id",
          "title",
          "cover",
          "position",
          "source",
          "score",
          "reason",
          "model_version"
        ],
        "title": "ItemDetailResponse"
      },
...
      "OnlineStatus": {
        "type": "string",
        "enum": [
          "online",
          "offline"
        ],
        "title": "OnlineStatus"
      },
      "OperationBatchRequest": {
        "properties": {
          "batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Batch Id"
          },
          "operation_type": {
            "type": "string",
            "enum": [
              "promote",
              "offline",
              "restore"
            ],
            "title": "Operation Type"
          },
          "targets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Targets"
          },
          "scope_type": {
            "type": "string",
            "enum": [
              "all",
              "user",
              "feed"
            ],
            "title": "Scope Type",
            "default": "all"
          },
          "scope_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Value"
          },
          "starts_at_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At Utc"
          },
          "ends_at_utc": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At Utc"
          },
          "priority": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Priority",
            "default": 0
          },
          "target_position": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Position"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason"
          },
          "semantics": {
            "type": "string",
            "const": "preflight_then_all_or_nothing_transaction",
            "title": "Semantics",
            "default": "preflight_then_all_or_nothing_transaction"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "batch_id",
          "operation_type",
          "targets",
          "starts_at_utc",
          "reason"
        ],
        "title": "OperationBatchRequest"
      },
      "OperationBatchResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Batch Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status"
          },
          "expected_state_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected State Version"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "batch_id",
          "status",
          "expected_state_version",
          "scheduled_at",
          "created_at"
        ],
        "title": "OperationBatchResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "AdminItem": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "heat": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Heat"
          },
          "online_status": {
            "type": "string",
            "enum": [
              "online",
              "offline"
            ],
            "title": "Online Status"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "state_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "State Version"
          },
          "cover": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "item_id",
          "title",
          "heat",
          "online_status",
          "updated_at",
          "state_version",
          "cover"
        ],
        "title": "AdminItem"
      },
      "AuditOperation": {
        "properties": {
          "operation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Operation Id"
          },
          "batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Batch Id"
          },
          "operator_id": {
            "type": "string",
            "format": "uuid",
            "title": "Operator Id"
          },
          "operator_role": {
            "type": "string",
            "enum": [
              "user",
              "operator_readonly",
              "operator",
              "admin"
            ],
            "title": "Operator Role"
          },
          "operation_type": {
            "type": "string",
            "enum": [
              "promote",
              "offline",
              "restore"
            ],
            "title": "Operation Type"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "targets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Targets"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "before_value": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Value"
          },
          "after_value": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Value"
          },
          "result": {
            "type": "string",
            "enum": [
              "succeeded",
              "failed"
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "effective_at": {
            "type": "string",
            "format": "date-time",
            "title": "Effective At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operation_id",
          "batch_id",
          "operator_id",
          "operator_role",
          "operation_type",
          "reason",
          "targets",
          "target",
          "before_value",
          "after_value",
          "result",
          "error",
          "effective_at"
        ],
        "title": "AuditOperation"
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "microlens_session"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
    "responses": {
      "Error": {
        "description": "Structured API error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Host-published public listener"
    }
  ],
...
}